docker_compose/simple_project/etl/requirements.txt
elasticsearch==8.4.1
psycopg2-binary==2.9.3
pydantic==1.10.2
python-dotenv==0.20.0
